// src/api/controller/testApi/testDTO.ts
export interface IPostTest {
  id: number;
  title: string;
  price: number;
  description: string;
  category: string;
  image: string;
}

export interface IPutTest {
  id: number;
  title: string;
  price: number;
  description: string;
  category: string;
  image: string;
}

export interface IDeleteTest {
  id: number;
}

export interface IGetTest {
  id: number;
}

export interface IResponseTest {
  id: number;
  title: string;
  price: number;
  description: string;
  category: string;
  image: string;
}